requirements.txt
booknlp
llama-index-llms-langchain
llama-index-readers-obsidian
llama-index
setuptools
spacy
transformers
wheel
pdfminer.six[image]
python-dotenv
claude-api-py
ffmpeg-python
groq
deepgram-sdk